0.7 error handling Ok

+++ rendu/includes/libft.h
@@ -30,7 +30,7 @@ void			ft_putnbr(int n);
 size_t			ft_strlen(const char *s);
 char			*ft_strcpy(char *dst, const char *src);
 char			*ft_strncpy(char *dst, const char *src, size_t len);
-int				ft_atoi(const char *str);
+long				ft_atoi(const char *str);
 int				ft_isalpha(int c);
 int				ft_isdigit(int c);
 int				ft_isalnum(int c);
